backend/pyproject.toml
[project]
name = "blog-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.14",
    "uvicorn>=0.34.3",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "notion-client>=2.2.1",
]
